fix: review issues with branch

+++ shopyo/modules/box__bizhelp/announce/view.py
@@ -1,6 +1,6 @@
 # from flask import render_template
-# from flask import url_for
-# from flask import redirect
+from flask import url_for
+from flask import redirect
 from flask import flash
 from flask import request
 
@@ -41,9 +41,7 @@ def add_check():
         form = AnnounceForm()
         if not form.validate_on_submit():
             flash_errors(form)
-            return redirect(
-                url_for("{}.dashboard".format(mhelp.info["module_name"]))
-            )
+            return mhelp.redirect_url("{}.dashboard".format(mhelp.info["module_name"]))
         toadd_announce = Announcement(
             content=form.content.data,
             title=form.title.data,
@@ -52,7 +50,7 @@ def add_check():
     return mhelp.redirect_url("{}.dashboard".format(mhelp.info["module_name"]))
 
 
-@module_blueprint.route("/list", methods=["GET", "POST"])
+@module_blueprint.route("/list", methods=["GET"])
 @login_required
 def list():
     context = mhelp.context()
@@ -65,19 +63,24 @@ def list():
 @login_required
 def delete_check(announce_id):
     announcement = Announcement.query.get(announce_id)
+    if announcement is None:
+        flash(notify_warning('Wrong announcement id'))
+        return mhelp.redirect_url(mhelp.method("list"))
     announcement.delete()
     return mhelp.redirect_url(mhelp.method("list"))
 
 
-@module_blueprint.route("/<announce_id>/edit", methods=["GET", "POST"])
+@module_blueprint.route("/<announce_id>/edit", methods=["GET"])
 @login_required
 def edit(announce_id):
     context = mhelp.context()
     announcement = Announcement.query.get(announce_id)
-
+    if announcement is None:
+        flash(notify_warning('Cannot find announcement id'))
+        return mhelp.redirect_url(mhelp.method("list"))
     form = AnnounceForm(obj=announcement)
 
-    context.update({"dir": dir, "announcement": announcement, "form": form})
+    context.update({"announcement": announcement, "form": form})
     return mhelp.render("edit.html", **context)
 
 
@@ -90,7 +93,7 @@ def edit_check(announce_id):
         form = AnnounceForm(obj=announcement)
         if not form.validate_on_submit():
             flash_errors(form)
-            return mhelp.redirect_url(mhelp.method("list"))
+            return mhelp.redirect_url(mhelp.method("edit"), announce_id=announce_id)
         form.populate_obj(announcement)
         announcement.update()
         flash(notify_success("Announcement updated!"))
